Working on zipped backups

+++ app/backup_generator.py
@@ -6,27 +6,29 @@ tmp_path = "tmp_saves"
 
 ## Function to copy save files to tmp folder requires a list with dicts as parameter "games"
 def copy_saves_toTmp(games):
-    for game in games:
-        if os.path.isdir(tmp_path + "/" + game["name"]):
-            print("Removing old {} save files".format(game["name"]))
-            shutil.rmtree(tmp_path + "/" + game["name"])
-            print("Done")
-        if os.path.isdir(game["path"]):
-            print("Copying {} save files".format(game["name"]))
-            shutil.copytree(game["path"], tmp_path + "/" + game["name"])
-            print("Done")
-
+    if isinstance(games, list) and isinstance(games[0], dict):
+        for game in games:
+            if os.path.isdir(tmp_path + "/" + game["name"]):
+                print("Removing old {} save files".format(game["name"]))
+                shutil.rmtree(tmp_path + "/" + game["name"])
+                print("Done")
+            if os.path.isdir(game["path"]):
+                print("Copying {} save files".format(game["name"]))
+                shutil.copytree(game["path"], tmp_path + "/" + game["name"])
+                print("Done")
+    else:
+        raise ValueError("Games parameter must be a list with dicts containing game attributtes")
 
 ## Function to make a zip archive # TODO: Replace shutil with a faster option,
 ## Add a index file with the paths for easier management
-def make_zip_file():
-    if os.path.isfile("ZippedBackups.zip"):
+def make_zip_file(filename="ZippedBackups.zip"):
+    if os.path.isfile(filename):
         print("Removig old zip")
-        os.remove("ZippedBackups.zip")
+        os.remove(filename)
         print("Done")
     if os.path.isdir(tmp_path):
         print("Making zip archive")
-        shutil.make_archive("ZippedBackups", "zip", root_dir=tmp_path)
+        shutil.make_archive(filename.replace(".zip", ""), "zip", root_dir=tmp_path)
         print("Done, Removing tmp_files")
         shutil.rmtree(tmp_path)
 
@@ -34,17 +36,27 @@ def make_zip_file():
 def read_zip_file(zipped="ZippedBackups.zip"):
     games = []
     zip_file = zipfile.ZipFile(zipped, "r")
-    for dir in zip_file.namelist():
-        if dir.endswith("/") == True and dir.count("/") == 1:
-            games.append(dir.replace("/", ""))
+    for folder in zip_file.namelist():
+        if folder.endswith("/") == True and folder.count("/") == 1 and folder != "./":
+            games.append(folder.replace("/", ""))
+    zip_file.close()
     return games
 
 
-def check_if_zipped(games, zipped="ZippedBackups.zip"):
-    zip_file = zipfile.ZipFile(zipped, "r")
+def check_if_zipped(games, zipped=read_zip_file()):
     for game in games:
-        if game in zip_file.namelist():
-            print(game)
+        if game["name"] in zipped:
+            print("Game:", game["name"])
+
+
+def append_to_zip(games, zipped="ZippedBackups.zip", overwrite=False):
+    if overwrite == False:
+        zip_file = zipfile.ZipFile(zipped, "r")
+        zipped_dirs = zip_file.namelist()
+        for folder in zipped_dirs:
+            if folder.endswith("/") == True and folder.count("/") == 1 and folder != "./":
+                if folder not in games:
+                    print(folder)
 
 
 def delete_from_list(games, name):
@@ -59,6 +71,7 @@ def delete_from_list(games, name):
 #copy_saves_toTmp(generate_games())
 #copy_saves_toTmp(pickle.load(open("games.pckl", "rb")))
 #make_zip_file()
-check_if_zipped(pickle.load(open("games.pckl", "rb")), "ZippedBackups.zip")
+#check_if_zipped(pickle.load(open("games.pckl", "rb")))
 #print(read_zip_file("ZippedBackups.zip"))
 #delete_from_list(pickle.load(open("games.pckl", "rb")))
+append_to_zip(pickle.load(open("games.pckl", "rb")))
